refactor(cdn): add reusable API utils, use in widget
