+++ cdn/src/widget.ts
@@ -1,22 +1,16 @@
-import type {
-  ApiErrorResponse,
-  PaymentInitiateInput,
-  PaymentInitiateResult,
-  PaymentQuoteInput,
-  PaymentQuoteResult,
-  PaymentStatus,
-  WalletAddressInfo,
-} from 'publisher-tools-api'
+import type { PaymentStatus } from 'publisher-tools-api'
 import { API_URL, APP_URL } from '@shared/defines'
 import type { WidgetProfile } from '@shared/types'
-import {
-  checkHrefFormat,
-  fromAmount,
-  sleep,
-  toWalletAddressUrl,
-} from '@shared/utils'
+import { sleep } from '@shared/utils'
 import { PaymentWidget } from '@tools/components'
-import { appendPaymentPointer, fetchProfile, getScriptParams } from './utils'
+import {
+  appendPaymentPointer,
+  fetchProfile,
+  fetchQuote,
+  getScriptParams,
+  getWallet,
+  initiatePayment,
+} from './utils'
 
 customElements.define('wm-payment-widget', PaymentWidget)
 
@@ -34,74 +28,21 @@ const drawWidget = (walletAddressUrl: string, profile: WidgetProfile) => {
   const frontendUrl = new URL('/tools/', getFrontendUrlOrigin()).href
   const element = document.createElement('wm-payment-widget')
   element.setController({
-    async getWallet(walletAddressUrl) {
-      walletAddressUrl = checkHrefFormat(toWalletAddressUrl(walletAddressUrl))
-
-      const url = new URL('/wallet', API_URL)
-      url.searchParams.set('walletAddress', walletAddressUrl)
-
-      const response = await fetch(url)
-      const data = await response.json()
-      if (!response.ok) {
-        throw new Error((data as ApiErrorResponse).error?.message)
-      }
-      return data as WalletAddressInfo
+    getWallet: (walletAddressUrl) => getWallet(API_URL, walletAddressUrl),
+    fetchQuote({ sender, receiver, amount }) {
+      const debitAmount = Number(amount)
+      return fetchQuote(API_URL, { sender, receiver, debitAmount })
     },
-    async fetchQuote({ sender, receiver, amount }) {
-      const url = new URL('/payment/quotes', API_URL)
-      const body: PaymentQuoteInput = {
+    initiatePayment({ sender, receiver, amount, note }) {
+      const debitAmount = Number(amount)
+      const redirectUrl = new URL('payment-confirmation', frontendUrl).href
+      return initiatePayment(API_URL, {
         sender,
         receiver,
-        debitAmount: Number(amount),
-      }
-      const res = await fetch(url, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(body),
-      })
-      const json: PaymentQuoteResult = await res.json()
-      if (res.ok && 'receiveAmount' in json) {
-        const debitAmount = fromAmount(json.debitAmount)
-        const receiveAmount = fromAmount(json.receiveAmount)
-        return { debitAmount, receiveAmount }
-      }
-
-      if (res.status === 400 && 'error' in json) {
-        const { error, minSendAmount } = json
-        return {
-          error,
-          ...(minSendAmount && { minSendAmount: fromAmount(minSendAmount) }),
-        }
-      } else {
-        return {
-          error: 'Failed to create payment. Please try a different amount.',
-        }
-      }
-    },
-    async initiatePayment({ sender, receiver, amount, note }) {
-      const url = new URL('/payment/initiate', API_URL).href
-      const body: PaymentInitiateInput = {
-        sender,
-        receiver,
-        debitAmount: Number(amount),
+        debitAmount,
         note,
-        redirectUrl: new URL('payment-confirmation', frontendUrl).href,
-      }
-      const res = await fetch(url, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(body),
+        redirectUrl,
       })
-      if (!res.ok) {
-        throw new Error(
-          `Failed to initiate payment. HTTP ${res.status} (${res.statusText})`,
-        )
-      }
-      const json: PaymentInitiateResult = await res.json()
-      return {
-        grantRedirectUrl: json.grantRedirectUrl,
-        paymentId: json.paymentId,
-      }
     },
     async *getStatus(paymentId, signal) {
       const url = new URL(`/payment/status/${paymentId}`, API_URL).href
